feat(revit): fully removes revit 2019 (#1858)
