+++ Core/Core/Kits/Applications.cs
@@ -133,91 +133,4 @@ namespace Speckle.Core.Kits
     }
 
   }
-
-  [Obsolete("Use HostApplications instead.", true)]
-  public static class Applications
-  {
-    public const string Rhino6 = "Rhino6";
-    public const string Rhino7 = "Rhino7";
-    public const string Grasshopper = "Grasshopper";
-    public const string Revit2019 = "Revit2019";
-    public const string Revit2020 = "Revit2020";
-    public const string Revit2021 = "Revit2021";
-    public const string Revit2022 = "Revit2022";
-    public const string Revit2023 = "Revit2023";
-    public const string DynamoSandbox = "DynamoSandbox";
-    public const string DynamoRevit = "DynamoRevit";
-    public const string DynamoRevit2021 = "DynamoRevit2021";
-    public const string DynamoRevit2022 = "DynamoRevit2022";
-    public const string DynamoRevit2023 = "DynamoRevit2023";
-    public const string Unity = "Unity";
-    public const string Excel = "Excel";
-    public const string GSA = "GSA";
-    public const string Civil2021 = "Civil2021";
-    public const string Civil2022 = "Civil2022";
-    public const string Civil2023 = "Civil2023";
-    public const string Autocad2021 = "AutoCAD2021";
-    public const string Autocad2022 = "AutoCAD2022";
-    public const string Autocad2023 = "AutoCAD2023";
-    public const string MicroStation = "MicroStation";
-    public const string OpenRoads = "OpenRoads";
-    public const string OpenRail = "OpenRail";
-    public const string OpenBuildings = "OpenBuildings";
-    public const string ETABSv18 = "ETABSv18";
-    public const string ETABSv19 = "ETABSv19";
-    public const string TeklaStructures2021 = "TeklaStructures2021";
-    public const string TeklaStructures2020 = "TeklaStructures2020";
-    public const string Archicad = "Archicad";
-    public const string Script = "Script";
-    public const string Other = "Other";
-    public const string All = "All";
-  }
-
-  // <summary>
-  // List of Host Applications with their versions 
-  // Do not change these - they must match the converter names!
-  // </summary>
-  [Obsolete("Use HostApplications.XXX.GetVersion() instead", true)]
-  public static class VersionedHostApplications
-  {
-    public const string Rhino6 = "Rhino6";
-    public const string Rhino7 = "Rhino7";
-    public const string Grasshopper6 = "Grasshopper6";
-    public const string Grasshopper7 = "Grasshopper7";
-    public const string Revit2019 = "Revit2019";
-    public const string Revit2020 = "Revit2020";
-    public const string Revit2021 = "Revit2021";
-    public const string Revit2022 = "Revit2022";
-    public const string Revit2023 = "Revit2023";
-    public const string Dxf = "Dxf";
-    public const string DynamoSandbox = "DynamoSandbox";
-    public const string DynamoRevit = "DynamoRevit";
-    public const string DynamoRevit2021 = "DynamoRevit2021";
-    public const string DynamoRevit2022 = "DynamoRevit2022";
-    public const string DynamoRevit2023 = "DynamoRevit2023";
-    public const string Unity = "Unity";
-    public const string Excel = "Excel";
-    public const string GSA = "GSA";
-    public const string Civil2021 = "Civil2021";
-    public const string Civil2022 = "Civil2022";
-    public const string Civil2023 = "Civil2023";
-    public const string Autocad2021 = "AutoCAD2021";
-    public const string Autocad2022 = "AutoCAD2022";
-    public const string Autocad2023 = "AutoCAD2023";
-    public const string MicroStation = "MicroStation";
-    public const string OpenRoads = "OpenRoads";
-    public const string OpenRail = "OpenRail";
-    public const string OpenBuildings = "OpenBuildings";
-    public const string ETABS = "ETABS";
-    public const string SAP2000 = "SAP2000";
-    public const string CSIBridge = "CSIBridge";
-    public const string SAFE = "SAFE";
-    public const string CSI = "CSI";
-    public const string TeklaStructures2021 = "TeklaStructures2021";
-    public const string TeklaStructures2020 = "TeklaStructures2020";
-    public const string Script = "Script";
-    public const string Other = "Other";
-    public const string All = "All";
-  }
-
 }
