+++ Objects/Converters/ConverterRevit/ConverterRevitShared/Partial Classes/ConvertTopography.cs
@@ -33,26 +33,12 @@ namespace Objects.Converter.Revit
           pts.Add(PointToNative(point));
         }
 
-#if !REVIT2019
-        facets.Capacity += (int) (displayMesh.faces.Count / 4f) * 3;
-        int j = 0;
-        while (j < displayMesh.faces.Count)
-        {
-          facets.Add(new PolymeshFacet(displayMesh.faces[j + 1], displayMesh.faces[j + 2], displayMesh.faces[j + 3]));
-          j += 4;
-        }
-#endif
       }
 
       if (docObj != null)
         Doc.Delete(docObj.Id);
 
-      TopographySurface revitSurface = null;
-#if !REVIT2019
-      revitSurface = TopographySurface.Create(Doc, pts, facets);
-#else
-      revitSurface = TopographySurface.Create(Doc, pts);
-#endif
+      var revitSurface = TopographySurface.Create(Doc, pts);
       if (speckleSurface is RevitTopography rt)
         SetInstanceParameters(revitSurface, rt);
